this was updated from home

+++ script.js
@@ -1,24 +1,31 @@
 window.addEventListener("load", () => {
     let service = new CarService("Car serivce");
+    hideRemove();
+    hideForm();
+    service.loadCars();
     const forms = document.querySelector("#form");
     const remove = document.querySelector("#form_remove");
 
     remove.addEventListener("submit", (e) => {
         e.preventDefault();
-        const formData = new FormData(remove);
-        console.log(formData.get("model"));
-        service.removeCar(formData.get("model"));
+        let submitter = e.submitter;
+        if (submitter.id != "remover") {
+            const formData = new FormData(remove);
+            service.removeCar(formData.get("model"));
+        }
+
 
     })
 
     forms.addEventListener("submit", (e) => {
         e.preventDefault();
-
-        const formData = new FormData(forms);
-
-        const car = new Car(formData.get("year"), formData.get("brand"), formData.get("model"), formData.get("km"));
-        console.log(car);
-        service.addCar(car);
+        let submitter = e.submitter;
+        if (submitter.id != "hidder") {
+            const formData = new FormData(forms);
+            const car = new Car(formData.get("year"), formData.get("brand"), formData.get("model"), formData.get("km"));
+            console.log(car);
+            service.addCar(car);
+        }
     })
 
 
@@ -43,6 +50,21 @@ class CarService {
         this.carArray = [];
         this.name = name;
     }
+    saveCars() {
+        window.localStorage.setItem("cars", JSON.stringify(this.carArray));
+    }
+
+    loadCars() {    
+        let tempcarArray = JSON.parse(window.localStorage.getItem("cars"));
+        if (tempcarArray != null) {
+            this.carArray = tempcarArray;
+            this.showCars();
+        }
+    }
+
+    removeCache() {
+        window.localStorage.clear();
+    }
 
     addCar(year, brand, model, km) {
         const car = new Car(year, brand, model, km);
@@ -51,6 +73,7 @@ class CarService {
 
     addCar(car) {
         this.carArray.push(car);
+        this.saveCars();
         this.showCars()
     }
 
@@ -61,15 +84,24 @@ class CarService {
                 index_catch = index;
             }
         })
-        this.carArray.slice(index_catch);
+        this.removeCar2(index_catch);
         this.showCars();
+        this.saveCars();
+    }
+    
+    removeCar2(index) {
+        delete this.carArray[index];
     }
 
     showCars() {
         this.clearCars();
         this.carArray.forEach((car, index) => {
-            this.createCard(car);
-        })
+            if (car != null) {
+                this.createCard(car);
+            } else {
+                this.removeCar2(index);
+            }
+        });
     }
 
     clearCars() {
@@ -107,18 +139,26 @@ class CarService {
     }
 }
 
+
 function showForm() {
-    const form = document.querySelector("#forms");
+    let form = document.querySelector("#forms");
+    hideRemove();
     form.style.display = "block";
 }
 
-function hideForm(object) {
-    const form = document.querySelector("#" + object);
+function hideForm() {
+    let form = document.querySelector("#forms");
+    form.style.display = "none";
+}
+
+function hideRemove() {
+    let form = document.querySelector("#remove");
     form.style.display = "none";
 }
 
 function showRemove() {
-    const form = document.querySelector("#remove");
+    let form = document.querySelector("#remove");
+    hideForm();
     form.style.display = "block";
 }
